Replace recipe resource dependencies with normal test dependencies
Since neither of the two recipe resource dependencies are used in any of the production code, I don't think there is much point in managing them that way. There is also a downside to it: They end up in the artifact deployed to Maven Central, which both adds to the artifacts size and the risk of them triggering CVE tools (like Nexus), even though they may never get used in the production code.

Fixes: #13