[83] Test that demonstrates GlobalDataFlow static variable issue
There is a find-nearest-method aspect of the `isSink` process that for

`bar(Constant.FOO)` sees `Constant` as the 'nearest object' to FOO.
This ultimately means it doesn't get seen as a sink.  But there's also
the 'pre-calculation' of edges that I'm a little concerned about.

Even so, very tempted to just modify findNearestMethod..  But is it
findNearestMethod or findNearestContainer...

I shouldn't have put the enum case in here.  I'll remove after a brief
mention of it..

 #83